QQuickWindow: Avoid shuffling around QSGNodes more than necessary
This optimization should help performance of the common cases
quite significantly, as entire subtrees of the scenegraph are no longer
shuffled around as much for no reason.

While the changes may look trivial, it's especially important to note that
changing children will likely need to recursively mark nodes in the
scenegraph backend, so the more expensive the subtree, the worse
the impact of this would have been - and the effects may have been quite
subtle (e.g. a lot of useless extra work, partial repaints not working, etc).

After this patch, the relevant tests in qmlbench change as follows:

    auto/changes/set_visible_in_long_tree_begin.qml  +25%
    auto/changes/set_visible_in_long_tree_middle.qml +15%
    auto/changes/set_visible_in_long_tree_end.qml    +0.4% (insignificant)

The edge cases for child changes (e.g. altering many children rather
than one at a time, or reshuffling the tree due to z order changes)
are still not very well optimized.

Improvement for those cases are left as an exercise for the reader, but
should at least work as well as they did before.

Change-Id: I556f6270f81579e88be53880e532f06ad18bf2f9
Reviewed-by: Andy Nichols <andy.nichols@qt.io>
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>
Reviewed-by: Laszlo Agocs <laszlo.agocs@qt.io>
Reviewed-by: Shawn Rutledge <shawn.rutledge@qt.io>