QQuickStates: Handle more types as result of when
When calling QQmlBinding::evaluate(), we get a generic QVariant back,
and no conversions from the evaluation result to bool are done at that
point. Consequently, we would need to mirror the engines conversion from
arbitrary values to bool.
Doing that manually is rather fragile, so we instead use the QMetaType
based overload, which already takes care of the conversion. That
requires making the overload public.
As qmlsc is already using that overload from within evaluate, it was not
affected by the  bug.

Fixes: QTBUG-108024
Change-Id: I3d5ee8f886ea6ada44fc8182300f3472cb1efc3c
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>
(cherry picked from commit dfdba3b12bc10a3e1da8cde99a8697969ac165ef)
Reviewed-by: Sami Shalayel <sami.shalayel@qt.io>