DA: don't ignore subtreeHoverEnabled
The current hover delivery logic in the DA was currently
only taking subtreeHoverEnabled into account for disabled
items. This meant that subtreeHoverEnabled was in practice
undermined by the DA, and resulted in far too many items in
the scene being visited.

Moreover, according to
tst_HoverHandler::hoverHandlerOnDisabledItem(), and
QTBUG-30801, we should always deliver hover events to
disabled items.

This patch will therefore change the test, so that we
always check if subtreeHoverEnabled is set before we
recurse into the child. Whether or not the item is enabled
or disabled is not relevant, and therefore removed.

This will ensure that we e.g don't visit all the delegate
items in a TableView for each pointer move, just because
TableView itself subscribes to hover (unless the
delegate is also subscribing to hover).

Note: This patch is only an optimization, and shouldn't
affect which items and handlers we end up actually
delivering hover events to. It will only stop the DA
from recursing into too many items.

Amends 831efa14e91cfa358e57a32197578d41c2ae1b24

Pick-to: 6.4
Change-Id: I6ddf963c3ac99ef54da08f64dc5660bc35a295e0
Reviewed-by: Shawn Rutledge <shawn.rutledge@qt.io>