Properly release Element data in the shadow tree
The suspicion is that in Qt 5 this was skipped because

1) releasing the generic allocation for any Element is done anyway due
   to destroying the m_elementAllocator, and

2) the only other thing releaseElement() did there was to release the
   extra data for QSGRenderNodes, which could thus be leaked, but
   render nodes are a rarely used advanced feature, and

3) upon closing a window, the scenegraph is torn down and the
   node-removed notification conveniently puts the Element data from
   the shadow nodes on the m_elementsToDelete queue (which is handled
   in the Renderer dtor); the same goes when dynamically removing a
   node at run time.

   So even if one cared about a few small leaks for apps with
   QSGRenderNode usage, to trigger this one would need to exercise
   some code path that created and released renderers, i.e. something
   with QSGLayer, and in a way that it does not lead to manipulating
   the scene graph. (so shader effects or Item.layer seem not to
   trigger this)

In Qt 6 calling releaseElement() is important for the most common
QSGGeometryNodes as well since there are additional resources to take
care of. And therefore the test code's grabToImage() (which internally
creates, renders, and then destroys a QSGLayer, and thus fires up and
then drops a full Renderer, while not touching the scenegraph itself
in any way) exhibits the problem much more clearly than in Qt 5.

Fixes: QTBUG-108026
Change-Id: Id60dfca7921efc2f3057c9cb8cedb238a1db2ee5
Reviewed-by: Christian Strømme <christian.stromme@qt.io>
Reviewed-by: Andy Nichols <andy.nichols@qt.io>
(cherry picked from commit 3b619f85aaaff997b5613f0add44098c8a3ca85c)