QQmlDataBlob: Take and return qreal for progress
QQmlDataBlob so far took and returned a quint8 for progress – however
the sole consumer of it expected a qreal in the [0;1] range.

progress returning an integer dates back all the way to
22a7d19d68bab2882fd9a6ebefd65152bbc1eaf5; the reason why this did not
create issues in practice is that the code path is only exercised when
downloading QML files over the network, which turned out to be a rather
uncommon use case.

Change-Id: I0140729955e393a19000adabe55f40fdd2caa04d
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>