ConstructionState: reduce size
We can reduce the size of the class by using QBiPointer to internally
store what previously has been two pointers and a flag.
The only draw-back is that we now have to manually manage the memory.
Mark the type as uncopieable to reduce the amount of boilerplate we
need; we still need to support moves as we store ConstructionState in a
vector.

Change-Id: I99329bbfda3f7147d475c1574a20a54719d2f7bd
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>