QQuickTableView: don't change index if tapping on the resize area
The current implementation would change the current index when
a single tap was done on the resize area between the cells.
But doing that was not ideal if the tap turned out to be
a double tap instead, since a double tap should only reset
the column size. The current solution would solve this by
using exclusive signals on the tap handler. But despite of
that, it would still look wrong that a tap on the resize
area, while the cursor has a resize appearance, changed the
index. Using exclusive signals also comes with a small signal
delay.

So, all in all, it's simply better to change the logic
so that the resize area is reserved for resizing only.
And then, to change the current index, the tap needs to
happen outside the resize area.

Change-Id: I1e7cb8d1031d1b86c14d777c322f1717fa5eb171
Reviewed-by: Shawn Rutledge <shawn.rutledge@qt.io>