Document dangers with parentChanged and visibleChanged
These signals get emitted by the QQuickItem destructor for the item under
destruction, but also for the visual child items, as they get removed
from the parent. Connected code that references parent objects might
crash.

We can't block those signal emissions as code relies on them to remove
pointers from data structures, and for items that are not also QObject-
children we have to emit the signals anyway as they will not get
destroyed, and their properties do change.

The best we can do is to document that it's potentially dangerous to
operate on visual parent objects in code connected to those signals.

Note that the visibleChanged signal does not get reliably emitted when
a parent in the visual hierarchy gets destroyed.

Pick-to: 6.4
Task-number: QTBUG-107850
Change-Id: I87913658818ec9491a4c4e799d9c553de44ea9c2
Reviewed-by: Shawn Rutledge <shawn.rutledge@qt.io>