QQuickTableView: detect if syncView topLeft has changed
It can happen that the top left cell in syncView changes
after a normal relayout. This is especially prone to
happen if the cells used to be large, but are suddenly
made much smaller, such that the top left cell ends
up outside the viewport.

This patch will detect this case, and ensure that the
top left is being recalculated when they are no
longer in sync.

Pick-to: 6.4 6.2
Change-Id: I80cad2859a35e0f3bc58f101696112be7c084175
Reviewed-by: Shawn Rutledge <shawn.rutledge@qt.io>