QQuickTableView: do a relayout whenever the view is resized
If you place a TableView in e.g a StackLayout, it will
have a zero size until it becomes the active view the
layout. If the delegate then has a binding that e.g
looks like:

implicitWidth: TableView.view.width

it will fail, since the width of the view is zero at time
of creation, but at the same time, never updated again
once the view becomes visible (and resized).

This patch will therefore ensure that we do a relayout
each time the view is resized, since having such a
dependency in the delegate (or in the columnWidthProvider)
must be expected to be quite common.

Change-Id: I92f4f51f6b6c634c2627716c41ea4971fb8d1653
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>
(cherry picked from commit 3cacdf288fa70e6f108b79a5752682329babf4c6)
Reviewed-by: Samuel Ghinet <samuel.ghinet@qt.io>