QQuickMaterialRipple: don't clip by default, leave it up to controls
The old button always specified a (clip)Radius of 2, so it never ran
into the bug where setting a radius of 0 would result in no clipping,
and a flat button with square corners would have a huge, perfectly
round background when e.g. hovered.

The ripple is a circle - that can be seen by looking at
QQuickMaterialRippleWaveNode::updateCurrentTime():

    rectNode->setRect(QRectF(0, 0, m_value, m_value));
    rectNode->setRadius(m_value / 2);

So, it doesn't make sense not to clip if the radius of the rectangle's
corners are 0; it always needs to clip a circle to fit in a rectangle.

However, for CheckBox, the ripple should always be round and not
clipped. So we should a) not clip by default and b) let each control
decide which behavior it wants.

Task-number: QTBUG-97993
Task-number: QTBUG-108052
Change-Id: I7c9f80151990874bd66ebb17304e65024f4384bd
Reviewed-by: Richard Moe Gustavsen <richard.gustavsen@qt.io>