Doc: Fix \brief for color QML value type
The following \target line was interpreted as being part of the
brief statement. Replace the \target with a \keyword which is
better for linking to a topic (instead of a subsection) and
move the command up in order not to clash with \brief.

Change-Id: I37c1058955f94ba3f9d36fbb078810b9ffdfcf40
Reviewed-by: Paul Wicking <paul.wicking@qt.io>
(cherry picked from commit 27d5f1b1335b31b4b26e882d853032fbf2b2df30)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>