QQuickTableView: factor out parts of the key handling into private function
Factor out most of the code in the keyPress() handler
that deals with changing the current index into a
private function. This will reduce a lot of d-pointer
indirection, but more importantly, simplify and prepare
the keyPress handler to be able to support the new
editDelegate API in later patches.

Change-Id: I2710a34261b4a28e45a03f3eb436d87dd1605c7a
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>