qqmljsscope: resolve enums using their aliases
Some enums happen to have an alias, e.g. when an enum is used as a flag, the
enum will exist in two versions, once as enum (e.g. Qt::MouseButton) and once
as a flag (e.g. Qt::MouseButtons). In this case, normally only the flag is
exposed to the qt metatype system and tools like qmltc will have troubles when
encountering the enum in signal parameters etc.

To solve this problem, QQmlJSScope::resolveEnums() will create a QQmlJSMetaEnum copy
for the alias in case the 'self'-scope already does not have an enum called
like the alias.

For the greater picture: this allows qmltc to compile signal handlers
for signals that take enums as argument without asserting, and also is
a step forward to compile the quick/text example with qmltc.

Task-number: QTBUG-107609
Change-Id: I8fcb48a15a2066fb2d2dc89fc26b8441a1a0f489
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>