Make tst_QQuickPinchHandler::scale easier to maintain
I kept forgetting why there were odd indented blocks, why it was letting
the QTouchEventSequence go out of scope and making a new one, and why
that resulted in an out-of-bounds pinch. It works well enough to just
drag the second touchpoint all the way down into the lower-right corner:
without maximumScale being set, it goes to 5.75, so we can verify that
setting maximumScale limits it to 4.

While we're at it:
- use QQuickTest::showView()
- better comments
- loop instead of copy-pasting blocks of code
- show the touchpoints
- slow it down enough to see them when cat. logging is turned on
- get rid of that weird opacity hack
- check pinchHandler's own scale() and activeScale() properties

Amends f289141524d9331bdafd9f88d95e0fa9b29ccdee

Pick-to: 6.2 6.4
Change-Id: I799c256286f4408f89c0502db33c9f676a2ce1b1
Reviewed-by: Richard Moe Gustavsen <richard.gustavsen@qt.io>