Visible signal is not emitted when item is removed or added to parent
Add tests that document this bug. When the visual parent of an item
is destroyed, then the item's parent becomes nullptr. Then the item is
no longer part of any visual hierarchy, and should no longer be visible
on screen.

However, the visibleChanged signal does not get emitted, or gets
emitted twice (and the visible property ends up set to true in the
end).

Note: the current behavior is the only reason why the "visible" test
in tst_qquickitem passes: the root item is created without a window,
and yet it, and its children, are expected to be visible.

Task-number: QTBUG-108213
Task-number: QTBUG-107850
Change-Id: I8dba24e7e0cf1899bf815d897b4466baaa86d31b
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>
Reviewed-by: Shawn Rutledge <shawn.rutledge@qt.io>