Remove broken assert in QQmlPropertyData::setOverrideIndex
The assert leads to strange crashes, as it compares if an overriding
and an overridden property share the same m_coreIndex and asserts when
equal.
This leads to crashes when overriding a function with a
property and vice-versa, as those can have an equal m_coreIndex but
still is a valid override.

Fixes: QTBUG-108627
Change-Id: I67f91d5fdb93603ef95b9a4557fd064edc24721d
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>
(cherry picked from commit d6961c09de36e15c57f29109edf5fbfef53ef4d4)
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>