Repair QQmlGadgetPtrWrapper's metaObject
We need two metaobjects: One with the PropertyAccessInStaticMetaCall
flagg and one without. The one without needs to be used with
QQmlGadgetPtrWrapper, since the wrapper wants to intercept metacalls.
The other one needs to be used when calling readOnGadget(). We can
accommodate this by just retaining the original metaobject. As it's
retrieved from a gadget type you should definitely be able to
readOnGadget() with it. The dynamic meta object, on the other hand, can
be lazily created when we actually metacall() through
QQmlGadgetPtrWrapper.

This relieves us of all the special casing around QQmlGadgetPtrWrapper
and makes it safe to use for anyone.

Fixes: QTBUG-108704
Fixes: QTBUG-111136
Fixes: QTBUG-111019
Change-Id: Icc01c81babaa9d1eca8d5ddfaf44d724a404db38
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>
(cherry picked from commit c4f0d07bda41ea3ea083d17c56e67a3b05a532bb)
Reviewed-by: Juha Vuolle <juha.vuolle@qt.io>