QQmlComponent/QQmlIncubator: Support C++ based types
QQmlComponent has a create overload which takes an Incubator. While
there is not really any work that could be done in an asynchronous with
a C++ based types (there are no sub-QML components, and there are no
bindings to set up), we still have to accommodate them in that overload
(otherwise we'd crash due to the missing compilation unit).

Add support in QQmlIncubator for this case, and call the newly introduce
function in QQmlComponent::create.

Task-number: QTBUG-97156
Change-Id: I64e3c6958117920dca214eee633424de94b492db
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>