Fix "white line on top and left side of screen" issue
Reproduced on Android, where devices sometimes have screen
sizes which are not evenly dividable by the device scale
factor (see QTBUG-87334 for examples).

Get the surface size for renderSceneGraph() directly from
the QPlatformWindow size. This avoids using the QWindow size,
which might have been rounded and can yield a surface size
which are off by a one or two of pixels.

Also change the "logicalSize" calculation to always use the
size passed to renderSceneGraph(), in order to avoid rounding
errors also here.

Fixes: QTBUG-87334
Change-Id: I23b5784305a876e1ef15ff6b01be09c6ce28409e
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>
Reviewed-by: Assam Boudjelthia <assam.boudjelthia@qt.io>