QQmlStandardPath: always use QUrl::fromLocalFile
It turns out that the changes done in 4e38a6a1948b69cab1
was too hasted. QStandardPaths are supposed to always
return local file paths, so converting them to URLs
using QUrl::fromLocalFile() was semantically correct.

The fact that the "local" file path on iOS was prefixed
with a scheme was not relevant - it should just be seen as
a part of the local file name. The iOS plugin that uses the
path even expects the URL to start with
"file:" ("file:assets-library://").

This patch will therefore partially revert
4e38a6a1948b69cab1, but keep the auto tests.
It turns out that it's the iOS plugin that loads the
image that needs to be fixed.

Change-Id: I1daaec7596bdbb8da5aec65c48bfd048f0c82d9d
Reviewed-by: Timur Pocheptsov <timur.pocheptsov@qt.io>
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>
(cherry picked from commit 9edd4becff099aae2a31f8e1af9f44393990a71d)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>