QQmlNotifier: Guard against broken meta objects when disconnecting
Some classes may be unable to produce a metaObject via the string
casting method we employ when invalidating the interceptor metaObject.
Ignore those.

Furthermore, qt_metacast() may produce a different object than the one
we called it on. Accept that. Finally, we don't have to check the
topmost QMetaObject because we already know it's invalid.

Fixes: QTBUG-108906
Change-Id: Id935d06beead4239381093f2f090d4a6a244034b
Reviewed-by: Zoltan Gera <zoltan.gera@qt.io>
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>