WorkerScript: Create ExecutionEngines on demand when needed
Creating an engine is rather expensive and as long as your script
doesn't run, we don't need one. Furthermore, this allows us to create
the engine in the same thread it will run in. Not only is this
stylistically better, but it will also let the engine check its stack
limits correctly on creation (once we do check stack limits).

Task-number: QTBUG-108182
Task-number: QTBUG-106875
Change-Id: Id0003b333c576b3bb603d8b71a5070211c7475bf
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>