QQuickOverlay: fix crash when a Drawer is destroyed
When a Drawer is deleted, the stack looks like this:
1 QQuickOverlayPrivate::removePopup
2 QQuickPopupPrivate::setWindow
3 QQuickPopup::setParentItem
4 QQuickPopup::~QQuickPopup
5 QQuickDrawer::~QQuickDrawer

It means that the QQuickPopup * we receive in
QQuickOverlayPrivate::removePopup() is not a QQuickDrawer * anymore.
Thus, the qobject_cast fails, we don't removeOne(), leave dangling
pointers in allDrawers, leading to a crash later in
QQuickOverlayPrivate::stackingOrderDrawers().

Store the Drawers as a list of QQuickPopup *. We downcast only in
QQuickOverlayPrivate::startDrag().

Pick-to: 6.4
Change-Id: I4840a18dec33426c58612db55937ea5988bf7650
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>